requirements.txt
numpy
matplotlib
python-chess